requirements.txt
loguru
boto3
pyalpm
python-libarchive
requests